calculate by double, more precise

+++ src/base/datatoimage.cpp
@@ -3,7 +3,7 @@
 #include "datatoimage.h"
 #include "glm.hpp"
 
-void DataToImage(const glm::vec3* data) {
+void DataToImage(const glm::dvec3* data) {
 
 	FILE *fp;
 	fp = fopen("../image/test.ppm", "w");
